script to calculate user monthly savings

+++ python_introduction/finance_calculator.py
@@ -2,5 +2,5 @@ monthly_income = int(input("Enter your monthly income: "))
 total_expenses = int(input("Enter your total monthly expenses: "))
 monthly_savings = monthly_income - total_expenses
 print("Your monthly savings are ", monthly_savins,".")
-projected_savings = monthly_savings * 12 + (monthly_davings * 12 *0.05))
+projected_savings = int(monthly_savings * 12 + (monthly_davings * 12 *0.05)))
 print("Projected savings after one year, with interest, is: ", projected_savins)
